Extract GenBank accession number from correspondence table

Extract GenBank accession number from correspondence table formatted with the same requirements for concatipede()

get_genbank_table(
  df = NULL,
  filename = NULL,
  writetable = F,
  out = "",
  excel.sheet = 1
)

Arguments

df
The user-defined correspondence table, as a data frame or equivalent. This is used only if no filename argument is provided.

filename
Filename of input correspondence table. Alternatively, if no filename is provided, the user can provide their own correspondence table as the df argument.

writetable
if TRUE save the Genbank table as excel file in the working directory

out
if writetable == T, the name to be attached to the excel filename

excel.sheet
specify what sheet from the excel spreadsheet you wanna read. Either a string (the name of a sheet), or an integer (the position of the sheet).

Value

Table with GenBank accession numbers