Build a template table with automatically matched sequence names

The algorithm used to match sequences across fasta files based on their names is outlined below.

auto_match_seqs(x, method = "lv", xlsx)

Arguments

x
A table (data frame or tibble) typically produced by concatipede_prepare. It must be of the same format as a table returned by this function: a first column called "name" followed by one column per fasta file. Those columns have the name of their corresponding fasta file, and they contain the names of the sequences in this file, with one sequence name per cell. The number of rows in the number of sequences of the fasta file with the most sequences, and the columns for the other fasta files are filled with NA for padding.

method
Method for string distance calculation. See ?stringdist::stringdist-metrics for details. Default is "lv".

xlsx
Optional, a path to use to save the output table as an Excel file.

Value

A table (tibble) with the same columns as x and with sequence names automatically matched across fasta files. Sequence names which did not have a best reciprocal match in other fasta files are appended to the end of the table, so that the output table columns contain all the unique sequence names present in the corresponding column of the input table. The first column, "name", contains a suggested name for the row (not guaranteed to be unique). If a path was provided to the xlsx argument, an Excel file is saved and the table is returned invisibly.

Details

Let's assume a situation with N fasta files, with each fasta file i having n_i sequence names. The problem of matching the names in the best possible way across the fasta files is similar to that of identifying homologous proteins across species, using e.g. reciprocal blast.

The algorithm steps are:

For each pair of fasta files, identify matching names using a reciprocal match approach: two names match if and only if they are their reciprocal best match.

Those matches across fasta files define a graph.

We identify sub-graphs such that (i) they contain at most one sequence name per fasta file and (ii) all nodes in a given sub-graph are fully connected (i.e., they are all their best reciprocal matches across any pair of fasta files).

Examples

xlsx_file <- system.file("extdata", "sequences-test-matching.xlsx",
                         package = "concatipede")
xlsx_template <- readxl::read_xlsx(xlsx_file)
auto_match_seqs(xlsx_template)

#> # A tibble: 405 x 5
#>    name   al18S_aligned_MA… al28S_aligned_MA… COI_aligned_Musc… ITS2_aligned_MA…
#>    <chr>  <chr>             <chr>             <chr>             <chr>           
#>  1 Macro… KT935502_Macrobi… KT935501_Macrobi… KT951668_Macrobi… KT935500_Macrob…
#>  2 Mesob… KX129793_Mesobio… KX129794_Mesobio… KX129796_Mesobio… KX129795_Mesobi…
#>  3 Tenui… KX810045_Tenuibi… KX810049_Tenuibi… KX810042_Tenuibi… KX810046_Tenuib…
#>  4 Macro… KY797265_Macrobi… KY797266_Macrobi… KY797267_Macrobi… KY797268_Macrob…
#>  5 Mesob… MF441488_Mesobio… MF441489_Mesobio… MF441491_Mesobio… MF441490_Mesobi…
#>  6 Param… MF568532_Paramac… MF568533_Paramac… MF568534_Paramac… MF568535_Parama…
#>  7 Macro… MH063922_Macrobi… MH063924_Macrobi… MH057764_Macrobi… MH063923_Macrob…
#>  8 Macro… MH063926_Macrobi… MH063935_Macrobi… MH057767_Macrobi… MH063931_Macrob…
#>  9 Mesob… MH197148_Mesobio… MH197265_Mesobio… MH195153_Mesobio… MH197156_Mesobi…
#> 10 Mesob… MH197149_Mesobio… MH197266_Mesobio… MH195154_Mesobio… MH197157_Mesobi…
#> # … with 395 more rows

if (FALSE) {
  auto_match_seqs(xlsx_template, xlsx = "my-automatic-output.xlsx")
}