Get the path to one of concatipede example files

Several example files are shipped with the concatipede package. This function facilitates the access to those files.

concatipede_example(example_file = NULL)

Arguments

example_file
Basename of the target example file. If NULL (the default), the basenames of the available files are listed.

Value

The full path to access the example file, or a list of available example files if no example_file argument was provided.

Details

COI_Macrobiotidae.fas

Example fasta file.

ITS2_Macrobiotidae.fas

Example fasta file.

LSU_Macrobiotidae.fas

Example fasta file.

SSU_Macrobiotidae.fas

Example fasta file.

sequences-test-matching.xlsx

This is an Excel file (extension .xlsx) typically used to test or demonstrate the automatic matching capabilities of the concatipede package. This file represents the Excel template that could be produced by concatipede_prepare after detecting the fasta files present in a working directory.

Macrobiotidae_seqnames.xlsx

This is an Excel file (extension .xlsx) that contains the correspondence table that can be used to concatenate the sequences contained in the example fasta files COI_Macrobiotidae.fas, ITS2_Macrobiotidae.fas, LSU_Macrobiotidae.fas, and SSU_Macrobiotidae.fas.

Examples

concatipede_example()

#> [1] "COI_Macrobiotidae.fas"        "ITS2_Macrobiotidae.fas"      
#> [3] "LSU_Macrobiotidae.fas"        "Macrobiotidae_seqnames.xlsx" 
#> [5] "sequences-test-matching.xlsx" "SSU_Macrobiotidae.fas"       

example <- concatipede_example("sequences-test-matching.xlsx")
if (requireNamespace("readxl")) {
  seqs <- readxl::read_xlsx(example)
  seqs
}

#> # A tibble: 330 × 5
#>     name al18S_aligned_MAF… al28S_aligned_MA… COI_aligned_Musc… ITS2_aligned_MA…
#>    <dbl> <chr>              <chr>             <chr>             <chr>           
#>  1     0 DQ839601_1_Macrob… FJ435751_Macrobi… AY598771_Dactylo… GQ403674_Parama…
#>  2     0 DQ839605_Paraacro… FJ435752_Macrobi… AY598772_Murrayo… GQ403675_Parama…
#>  3     0 EU038078_Paramacr… FJ435753_Macrobi… AY598773_Macrobi… GQ403676_Parama…
#>  4     0 EU038080_Paramacr… FJ435754_Macrobi… AY598774_Macrobi… GQ403678_Parama…
#>  5     0 EU038081_Paramacr… FJ435755_Macrobi… AY598775_Macrobi… GQ403679_Parama…
#>  6     0 EU266926_Macrobio… FJ435756_Paramac… AY598776_Macrobi… GQ403680_Macrob…
#>  7     0 EU266928_Mesobiot… FJ435757_Paramac… AY598777_Macrobi… HM150647_Macrob…
#>  8     0 EU266932_Minibiot… FJ435758_Minibio… EU244597_Paramac… KT935500_Macrob…
#>  9     0 EU266933_Minibiot… FJ435759_Minibio… EU244598_Paramac… KX129795_Mesobi…
#> 10     0 EU266934_Minibiot… FJ435760_Minibio… EU244599_Paramac… KX810010_Macrob…
#> # … with 320 more rows